chore(get): factor get impl
